Clear Swiftlint warning and separate SnapshotTests in multiple test cases
